requirements.txt
anthropic~=0.36.0
streamlit~=1.39.0
langchain~=0.3.3
SQLAlchemy~=2.0.35
pypdf~=5.0.1
python-docx~=1.1.2